+++ test/vector/collinear_with_test.rb
@@ -4,24 +4,24 @@ require 'geometry'
 class CollinearWithTest < Minitest::Test
   include Geometry
 
-  def test_vectors_are_collinear
-    vector1 = Vector.new(1, 2)
-    vector2 = Vector.new(2, 4)
+  def test_gvectors_are_collinear
+    gvector1 = GVector.new(1, 2)
+    gvector2 = GVector.new(2, 4)
 
-    assert vector1.collinear_with?(vector2)
+    assert gvector1.collinear_with?(gvector2)
   end
 
-  def test_vectors_are_not_collinear
-    vector1 = Vector.new(1, 2)
-    vector2 = Vector.new(1, 1)
+  def test_gvectors_are_not_collinear
+    gvector1 = GVector.new(1, 2)
+    gvector2 = GVector.new(1, 1)
 
-    assert ! vector1.collinear_with?(vector2)
+    assert ! gvector1.collinear_with?(gvector2)
   end
 
-  def test_vectors_are_oppositely_directed
-    vector1 = Vector.new(2, 2)
-    vector2 = Vector.new(-2, -2)
+  def test_gvectors_are_oppositely_directed
+    gvector1 = GVector.new(2, 2)
+    gvector2 = GVector.new(-2, -2)
 
-    assert vector1.collinear_with?(vector2)
+    assert gvector1.collinear_with?(gvector2)
   end
 end