+++ CHANGELOG.md
@@ -5,6 +5,6 @@ Changelog
 
 * Added `Line#distance_to(point)`, `Segment#distance_to(point)`
 * Added `Polygon#area`
-* Added `Point.advance_by(vector)`
+* Added `Point.advance_by(gvector)`
 * Switched to Ruby 2.0 as a default Ruby version to support
 * Switched from obsolete Test::Unit to Minitest